requirements.txt
python-dotenv>=1.0.1
requests>=2.28.2
setuptools>=57.4.0
Pillow>=9.5.0
pyelftools>=0.31
unicorn>=2.0.1.post1
windows>=0.1.1
pytz>=2024.1